Image blurring
can find python codes for this too. However, I am unsure if this can be done on pixel data that we have. Conceptually it is possible- idea being maths (Kernel) used for blurring image can be applied for pixel data post correlation analyses which is to be be performed in points 4 and 5. So instead of just using top highly correlated scores- we can modify the data such that rest of the attributes are blurred out.